Logging: write both stdout and stderr output to the same file
It gets harder to debug otherwise, since we have applications writing to both stdout and stderr and we want the debug experience to be reasonable. This was surprisingly hard, because of our very silly "open the file, write a line and close it" approach to logging. The latter is admittedly lame, but it has one significant advantage to just keeping the file open: it handles date rollovers in an extremely nice and pleasant way. Yes, the performance and CPU usage is clearly not the best with this; we might have to change it later on, but... it has served us decently for years, with reasonable amount of volume.

The locking is naive, but good enough for now.